+++ packages/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @mateus-ignite-ui/react
 
+## 4.1.0
+
+### Minor Changes
+
+- add new component - Tooltip
+
 ## 4.0.0
 
 ### Major Changes