+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @mateus-ignite-ui/tokens
 
+## 6.15.0
+
+### Minor Changes
+
+- add new component - Tooltip
+
 ## 6.14.0
 
 ### Minor Changes